--- plancontext/context_test.go
@@ -1,16 +1,0 @@
-package plancontext
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-)
-
-func TestContext(t *testing.T) {
-	ctx := New()
-
-	secret := ctx.Secrets.New("test")
-	get, err := ctx.Secrets.FromValue(secret.MarshalCUE())
-	require.NoError(t, err)
-	require.Equal(t, "test", get.PlainText())
-}
